xonai-grafana/requirements.txt
boto3==1.28.62
fastapi==0.103.2
prometheus-api-client==0.5.4
retrying==1.3.4
uvicorn==0.23.2
botocore==1.31.62
jmespath~=1.0.1
starlette~=0.27.0 # FastAPI requires this version
python-dateutil~=2.8.2
requests~=2.31.0
pydantic~=2.4.2
setuptools~=68.2.2
databricks-sdk~=0.10.0